Refactor remove weird QuickReservation export
