+++ apps/ui/components/reservation-unit/index.ts
@@ -1,4 +1,3 @@
-// export { QuickReservation } from "./QuickReservation"; Don't export due to circular dependency issues
 export { ReservationInfoSection } from "./ReservationInfoSection";
 export { ReservationUnitCalendarSection } from "./ReservationUnitCalendarSection";
 export { Head } from "./Head";
